refactor(customer-response): 잘못된 패키지 배치 수정. request -> response
