+++ src/test/java/woowacourse/shoppingcart/acceptance/CustomerAcceptanceTest.java
@@ -18,9 +18,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.test.context.jdbc.Sql;
 import woowacourse.shoppingcart.ui.dto.request.CustomerDeleteRequest;
 import woowacourse.shoppingcart.ui.dto.request.CustomerRequest;
-import woowacourse.shoppingcart.ui.dto.request.CustomerResponse;
 import woowacourse.shoppingcart.ui.dto.request.CustomerUpdatePasswordRequest;
 import woowacourse.shoppingcart.ui.dto.request.CustomerUpdateProfileRequest;
+import woowacourse.shoppingcart.ui.dto.response.CustomerResponse;
 import woowacourse.shoppingcart.ui.dto.response.ExceptionResponse;
 
 @DisplayName("회원 관련 기능")
